move scanmode and scantype dataclass in dataclass.py
